feat: mentor access > updated adding/updating user to sync with rbac & toggle manual/auto mode > tests coverage
