+++ __tests__/update-access.test.tsx
@@ -13,8 +13,6 @@ const mockUseUsername = vi.fn();
 const mockUseGetMentorSettingsQuery = vi.fn();
 const mockUsePlatformUsersQuery = vi.fn();
 const mockUpdateMentorAccess = vi.fn();
-const mockGetRbacPermissions = vi.fn();
-const mockDispatch = vi.fn();
 const mockRbacPermissions = vi.fn();
 const mockCheckRbacPermission = vi.fn();
 
@@ -30,18 +28,15 @@ vi.mock('@iblai/iblai-js/data-layer', () => ({
   useGetMentorSettingsQuery: (...args: unknown[]) => mockUseGetMentorSettingsQuery(...args),
   usePlatformUsersQuery: (...args: unknown[]) => mockUsePlatformUsersQuery(...args),
   useUpdateRbacMentorAccessMutation: () => [mockUpdateMentorAccess],
-  useGetRbacPermissionsMutation: () => [mockGetRbacPermissions],
   isPoliciesResponse: () => false,
 }));
 
 vi.mock('@/lib/hooks', () => ({
   useAppSelector: () => mockRbacPermissions(),
-  useAppDispatch: () => mockDispatch,
 }));
 
 vi.mock('@/features/rbac/rbac-slice', () => ({
   selectRbacPermissions: 'selectRbacPermissions',
-  updateRbacPermissions: (payload: unknown) => ({ type: 'rbac/updateRbacPermissions', payload }),
 }));
 
 vi.mock('@/hoc/withPermissions', () => ({
@@ -100,43 +95,7 @@ describe('RoleAccessPanel', () => {
     });
     mockRbacPermissions.mockReturnValue({});
     mockCheckRbacPermission.mockReturnValue(true);
-    mockGetRbacPermissions.mockReturnValue({ unwrap: () => Promise.resolve({}) });
     mockUpdateMentorAccess.mockReturnValue({ unwrap: () => Promise.resolve({}) });
-  });
-
-  /* ---------- RBAC permission fetch on mount ---------- */
-
-  it('fetches RBAC permissions on mount', async () => {
-    setup();
-
-    await waitFor(() => {
-      expect(mockGetRbacPermissions).toHaveBeenCalledWith({
-        requestBody: {
-          platform_key: 'test-tenant',
-          resources: ['/users/'],
-        },
-      });
-    });
-  });
-
-  it('dispatches updateRbacPermissions after fetch', async () => {
-    const permResult = { '/users/': { list: true } };
-    mockGetRbacPermissions.mockReturnValue({ unwrap: () => Promise.resolve(permResult) });
-
-    setup();
-
-    await waitFor(() => {
-      expect(mockDispatch).toHaveBeenCalledWith({
-        type: 'rbac/updateRbacPermissions',
-        payload: { ...permResult },
-      });
-    });
-  });
-
-  it('does not fetch permissions when tenantKey is missing', () => {
-    mockUseParams.mockReturnValue({ tenantKey: undefined, mentorId: 'test-mentor' });
-    setup();
-    expect(mockGetRbacPermissions).not.toHaveBeenCalled();
   });
 
   /* ---------- Assigned users display ---------- */
@@ -196,14 +155,6 @@ describe('RoleAccessPanel', () => {
       expect(screen.queryByLabelText('Select input type')).not.toBeInTheDocument();
     });
 
-    it('shows description for search mode', () => {
-      setup();
-
-      expect(
-        screen.getByText('Search by name, username, or email to grant this role to additional users.'),
-      ).toBeInTheDocument();
-    });
-
     it('adds user via search and calls mutation with users_to_add', async () => {
       mockUsePlatformUsersQuery.mockReturnValue({
         data: {
@@ -248,14 +199,6 @@ describe('RoleAccessPanel', () => {
       expect(screen.getByPlaceholderText('user@example.com')).toBeInTheDocument();
       expect(screen.getByRole('button', { name: /add entry/i })).toBeInTheDocument();
       expect(screen.queryByPlaceholderText('Search by name, username, or email')).not.toBeInTheDocument();
-    });
-
-    it('shows description for manual mode', () => {
-      setup();
-
-      expect(
-        screen.getByText('Enter a username or email to grant this role to additional users.'),
-      ).toBeInTheDocument();
     });
 
     it('stages entries via Enter key', async () => {
